+++ npf/output/grapher.py
@@ -15,10 +15,17 @@ from npf.output.statistics import Statistics
 from npf.output.transform.pandas import to_pandas
 from npf.output.transform.combine_variables import combine_variables
 from npf.output.transform.result_as_variable import result_as_variable
-from npf.tests import variable
+from npf.types import units
 from npf.output.notebook.notebook import prepare_notebook_export
 from npf.output.web.web import prepare_web_export
 from npf.types.units import *
+from npf.types.units import is_numeric
+import npf.types.units
+from npf.types.units import get_bool
+from npf.types.units import is_bool
+from npf.types.units import get_numeric
+import npf.types.units
+from npf.types.units import numericable
 if sys.version_info < (3, 7):
     from orderedset import OrderedSet
 else:
@@ -35,7 +42,7 @@ from math import log,pow
 from npf.types import dataset
 from npf.types.series import Series
 from npf.types.dataset import Run, XYEB, AllXYEB, group_val, var_divider, mask_from_filter
-from npf.tests.variable import is_log, is_numeric, get_numeric, numericable, get_bool, is_bool
+from npf.tests.variable import is_log
 from npf.tests.section import SectionVariable
 from npf.tests.build import Build
 from npf.output.graph.graph_choice import decide_graph_type
@@ -919,7 +926,7 @@ class Grapher:
                 matched = False
                 for k in data_types.keys():
                     if re.match(result_type, k):
-                        if variable.is_numeric(n_cols):
+                        if npf.types.units.is_numeric(n_cols):
                             n_cols = get_numeric(n_cols)
                             subplot_legend_titles = [self.var_name("result",result_type=result_type)]
                         else:
@@ -1753,8 +1760,8 @@ class Grapher:
                 for yi in range(len(xdata[i][2])):
                     x.append(np.mean(xdata[i][2][yi][2]))
             if not all_num(x):
-                if variable.numericable(x):
-                    ax = [variable.get_numeric(v) for i, v in enumerate(x)]
+                if units.numericable(x):
+                    ax = [npf.types.units.get_numeric(v) for i, v in enumerate(x)]
                 else:
                     ax = np.arange(len(x)) + 0.5 + i
             else:
@@ -1944,6 +1951,6 @@ class Grapher:
         axis.yname = yname
 
         if yticks:
-            ticks = [variable.get_numeric(parseUnit(y)) for y in yticks.split('+')]
+            ticks = [npf.types.units.get_numeric(parseUnit(y)) for y in yticks.split('+')]
             plt.yticks(ticks)
