+++ npf/output/transform/result_as_variable.py
@@ -1,7 +1,8 @@
 from collections import OrderedDict
 from ordered_set import OrderedSet
 
-from npf.tests.variable import get_numeric, numericable
+from npf.types.units import numericable
+from npf.types.units import get_numeric
 
 
 def result_as_variable(series, result_types, var_name, vars_values):
