+++ npf/tests/repository.py
@@ -15,9 +15,11 @@ import npf
 from npf import osutils
 import npf.globals
 from npf.tests.build import Build
-from npf.tests.variable import is_numeric, VariableFactory
+from npf.tests.variable import VariableFactory
 
 import git
+
+from npf.types.units import is_numeric
 
 repo_variables = ['name', 'branch', 'configure', 'url', 'method', 'parent', 'tags', 'make', 'version', 'clean', 'build_info',
                   'bin_folder', 'bin_name', 'env']
